Touch ups and moved PendingApprovals folder into the serviceScreen folder
